+++ rubato/utils/rendering/__init__.py
@@ -1,5 +1,5 @@
 """This module contains rendering utilities"""
 from .surface import Surface
 from .font import Font
-from .draw import Draw, DrawTask
+from .draw import Draw
 from .camera import Camera