+++ CHANGELOG.md
@@ -19,6 +19,11 @@
 
 ### Breaking Changes
 
+-   Hidden a LOT of methods that weren't meant to be touched by users.
+-   Renamed `surface.get_size()` to `surface.size_scaled()` and `surface.get_size_raw()` to `surface.size()`.
+-   Removed `Game.camera`. Good code should use the scene's camera directly or `Game.current().camera`.
+-   Made `Game.current()` a classmethod instead of a property. This is to make it more obvious that it is read-only.
+
 ### Added
 
 ### Changed